requirements.txt
django>=5.0,<=5.1.4
djangorestframework>=3.15.0,<=3.15.2
python-decouple<=3.8
djoser>=2.2.3,<=2.3.1
djangorestframework_simplejwt<=5.3.1
drf_yasg<=1.21.8
django-cors-headers<=4.6.0
pillow>=10.3.0,<=11.0.0
psycopg2
cloudinary
dj_database_url